porting shufflenetv2
